% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WpoisCI.R
\name{WpoisCI_lower}
\alias{WpoisCI_lower}
\title{An Admissible Exact One-sided Lower Confidence Interval for Poisson Mean}
\usage{
WpoisCI_lower(x, conf.level = 0.95, details = FALSE)
}
\arguments{
\item{x}{the sample or the observed point.}

\item{conf.level}{confidence level. The default is 0.95.}

\item{details}{TRUE/FALSE, can be abbreviated. To choose whether to compute
the confidence intervals for all the sample points. Default is FALSE.}
}
\value{
a list which contain the one-sided lower confidence interval.
}
\description{
The 1-alpha Clopper-Pearson type lower interval for the Poisson
mean.
}
\examples{
WpoisCI_lower(1)
WpoisCI_lower(3,details = TRUE)
}
\references{
Garwood, F. (1936). Fiducial Limits for the Poisson
Distribution. "Biometrika" 28:	437-442.
}
